Add interactive functions to mark completed task and delete completed tasks
