+++ src/modules/toDoListHelper.js
@@ -1,0 +1,70 @@
+class ToDoListHelper {
+  static checkBoxEventHandler = (toDoList) => {
+    const checkBox = document.querySelectorAll('input[type=checkbox]');
+    checkBox.forEach((item) => {
+      item.addEventListener('change', (event) => {
+        const task = event.target.parentNode.querySelector('.item-description');
+        task.classList.toggle('strike-line');
+        toDoList.updateTaskStatus(parseInt(event.target.id, 10) - 1);
+      });
+    });
+  };
+
+  static deleteButtonEventHandler = (toDoList) => {
+    const deleteButtons = document.querySelectorAll('.delete');
+    deleteButtons.forEach((item) => {
+      item.addEventListener('click', (event) => {
+        toDoList.deleteTaskFromList(false, parseInt(event.target.getAttribute('data-index'), 10));
+      });
+    });
+  };
+
+  static editElipsisEventHandler = (toDoList) => {
+    const descriptions = document.querySelectorAll('.item-description');
+
+    descriptions.forEach((item) => item.addEventListener('keypress', (event) => {
+      if (event.key === 'Enter') {
+        event.preventDefault();
+        this.editDescriptionHandler(event, toDoList);
+      }
+    }));
+
+    descriptions.forEach((item) => item.addEventListener('change', (event) => {
+      this.editDescriptionHandler(event, toDoList);
+    }));
+  };
+
+  static editDescriptionHandler = (event, toDoList) => {
+    const id = parseInt(event.target.getAttribute('data-pos'), 10);
+    const data = event.target.value;
+    toDoList.editDescription({ id, data });
+    event.target.disabled = true;
+    event.target.parentNode.querySelector('.modify').classList.toggle('d-none');
+  }
+
+  static editTextEventHandler = () => {
+    const description = document.querySelectorAll('.modify');
+    const descriptions = document.querySelectorAll('.item-description');
+
+    description.forEach((item) => {
+      item.addEventListener('click', (event) => {
+        descriptions.forEach((e) => {
+          e.disabled = true;
+          e.parentNode.classList.remove('highlight');
+          e.parentNode.querySelector('.modify').classList.remove('d-none');
+          e.parentNode.querySelector('.fa-trash').classList.add('d-none');
+        });
+
+        event.target.parentNode.classList.add('highlight');
+        const itemDescription = event.target.parentNode.querySelector('.item-description');
+        const deleteButton = event.target.parentNode.querySelector('.fa-trash');
+        itemDescription.disabled = false;
+        itemDescription.focus();
+        event.target.classList.toggle('d-none');
+        deleteButton.classList.toggle('d-none');
+      });
+    });
+  }
+}
+
+export default ToDoListHelper;
